Adds a backend to time between received key presses
The business logic is currently a loop to read events from the channel.
There will be also need for a logic to send the next paragraph to train
on.

Logic:
===
Start -> random stream of paragraphs (attempt?)
    words are randomly picked from which corpus dictionary?
Session -> timed session of 30mn (or other per day)
For each key pressed we need to make progress on the paragraph with
green/red
The paragraph is finished when the last key is pressed
The stats can be computed on the paragraph (and can be replayed?)
we can dump them in a db.